Ensure that the descriptions are faithfully passed to trac
This removes the previous wrapping of the descriptions, which was
presumably aimed more at keeping them sane for the terminal display
than for trac. Since the trac form of the tickets is more important
(and there are other elements which might be too wide anyway) remove
this wrapping so that more advanced trac formatting can be used in
the ticket descriptions.

Change-Id: Ie9933512f3fa3ba9163a1409e4300a2c6eac050b