Decouple prep for the virtual competition from the event itself
The planning and setup for the virtual competition can, should (and
likely will) happen a long time (~months) in adavance of the event
itself. Previously the tickets did not account for this, forcing
long running tickets which would have no meaningful update and
could not be closed and likely to have parts forgotten. This change
separates out the parts which are more likely to be done together
as well as connecting up some missing dependencies.