Separate out the teams/volunteers reminder emails for Kickstart
This decouples getting signups for the year as a whole from sending
the reminder emails for the first (public) event of the year. The
signup tasks can, should (and likely will) happen long before the
Kickstart event, yet the reminder email is something we'll send out
only a few days before.

Previously the coupling of these meant that we had tasks which were
forced to stay open for a long time, not getting updates, creating
confusion and increasing the chances that something would be missed.

This also adds in some overlooked dependencies.